requirements.txt
streamlit==1.36.0
requests>=2.31.0
pillow>=10.3.0
numpy>=1.26.0
opencv-python>=4.9.0.80
pandas>=2.2.2
pyarrow>=16.1.0
h5py>=3.11.0
tqdm>=4.66.4
python-dotenv>=1.0.1
imagehash>=4.3.1
transformers>=4.42.3
sentencepiece>=0.2.0
torch>=2.3.1
scikit-image>=0.24.0
urllib3>=2.2.2
beautifulsoup4>=4.12.3
lxml>=5.2.2
